frontend/requirements.txt
SQLAlchemy
streamlit
pandas
pymysql
PyYAML
numpy
streamlit_authenticator